Improve sensinact-test manifest generation
To be able to use sensinact-test the manifest of the bundle being built was needed. All the used ones were created outside of the test context and next added. Now the maven-bundle-plugin is used to generate the appropriate manifest file by the use of the 'manifest' goal before the test execution