Fix subscription and agent callbacks inconsistencies
Subscription and agent callbacks mechanisms reorganization to make them more consistent. The callback system of an agent was not related to the one associated to the subscription. Define the AbstractMidCallback type to factorize code and use the same underlying mechanism

The sent callback messages are now encapsulated into a JSONArray

Update all the data structures that were using the callbacks mechanisms