Fix multiple patterns applying on one device error
The access tree data structure as well as the database allow to define multiple patterns applying on a single device, but the DAO layer and more particulary the ObjectDAO data structure was not allowing to use such rules. Moreover when multiple patterns applies the ObjectDAO find method return null as only one ObjectEntity is expected. To fix the bug, define the returned type of the ObjectDAO'find methods as a List of ObjectEntity