Fix multi-pattern service provider matching
When a service provider identifier was matching multiple patterns defined in the database, an NullPointerException was triggered because of a null set of accessible methods found for the service provider (one of its held resource). Fix ObjectDAO to allow to retrieve multiple entries in the database. Fix search mechanism in PathTree. Add entries in initial database and definitino SQL script to test the mechanism. Add test