Fix SocketEndpoint initial connection error message
The SocketManager pre-connected the SocketEndpoints before calling the official connect method offered by the Core service - It resulted in an error message because of an invalid variables initialization chronology. The pre-connection is not done anymore. The delay of the official connection has been increased

Use optional methods set to parse the JSON data structure describing an agent to avoid JSONExceptions