SnaAgents management through remote Cores
The creation of a remote agent was potentially in error because of the asynchronous connection of the client and the server of a single SocketEndpoint. Morever the possibility of a null Snafilter was not handled correctly.

The SocketEndpoint wait now for connection of both its client and its server to be defined as 'connected'

A null SnaFilter is now handled correctly

Add test to check the appropriate use of an SnaAgent through differents sensiNact instances connected using SocketEndpoints