Refactor of the MQTT bridge
Separation of the sensinact-mosquitto in mqtt-device and smart-topic-device.
The mqtt-device is now an abstract bundle that can be used to create others MQTT bridges.
The smart-topic-device is used to create devices at runtime by just providing a configuration file.
Update of the POMs, the integration tests and of the distribution generator.