Introduce LoginEndpoint
Login process and calls to access methods were using the same endpoint. To avoid unecessary login process and to make NorthboundEndpoint reactivation easier, the LoginEndpoint type has been introduced; it allows to login into the platform and to obtain a valid session token, and it allows to reactivate an existing NorthboundEndpoint by transmitting it the token of the session it is linked to.