Make StackEngine share threads
The StackEngine currently creates one thread (and potentially one timer thread) per instance. This is problematic as each service provider has a Stack Engine, meaning the scaling is poor. This is a minimal fix to share an executor between the instances, but a better threading/broker model is probably needed in the long term.

Signed-off-by: Tim Ward <timothyjward@apache.org>