Contribute an implementation of the OGC Sensorthings MQTT API
In addition to its REST API the OGC Sensorthings specification has an MQTT binding. This reuses the same DTOs and is therefore relatively simple for us to support.

Note that the way in which the specification works requires the MQTT broker to be hosted in the gateway (so that we can spy on thesubscriptions for all the clients). We therefore repackage the Moquette broker.

This initial implementation covers the base collections, selection by id, and the  filter. The remaining property selection and entity sets should be trivial to add later.

Signed-off-by: Tim Ward <timothyjward@apache.org>