SensorThings MQTT: Queue notifications to keep their order
Integration tests were failing because sensiNact notifications and their
handling were done in (at least) 2 threads, meaning that the order of
sending of notification couldn't be guaranteed.

This commit uses a single thread that handles the publication of
messages in the same order as the notifications, still using multiple
threads to handle notifications and payload generation.