Implement the expand query for SensorThings northbound
This commit adds a backing implementation for the expand query for the SensorThings REST API. This is a pervasive change which requires us to carry context throughout the query so as to attach additional data structures based on the filter content.

* All queries now use full ProviderSnapshot instances so that we don't have to re-query for other resource data when filling in expansions
* Limited bulk data is included in expansions to prevent the JSON size exploding
* Data context is mapped by id to ensure we can expand values in lists
* All context is carried outside the DTO to avoid unnecessary fields appearing in the JSON output

Note that we require that all ids are able to be turned into strings. This is currently true throughout the gateway.

Signed-off-by: Tim Ward <timothyjward@apache.org>