Add a default TimedValue implementation
The TimedValue interface is an important part of the API, but I noticed an increasing number of _minimal_ implementations that duplicated one another. This isn't a good sign, and means that we should provide a default implementation. As TimedValue is supposed to be immutable and DTO-like this seems like a good time to move to Java 17 and implement the default as a record type. This keeps the API package clean and simple, while allowing us to continue using the TimedValue interface in the API. Where people want or need to create their own TimedValue implementations they can, but the record implementation should serve for most use cases.

Signed-off-by: Tim Ward <timothyjward@apache.org>