Remove the webclient
As discussed in issue #534 the webclient is being removed as it does not form part of the core gateway and there is no long-term commitment to maintaining it at this time. The code will remain in the repository history so that it can be revived if anyone is interested in it.

Fixes #534

Signed-off-by: Tim Ward <timothyjward@apache.org>