Initial version of a Southbound SensorThings connector
This commit adds support for mirroring SensorThings devices into the sensiNact digital twin. It includes a dynamic model that is used to represent the devices, a polling loop which queries for the latest data, and Records which map the detected updates into the defined model.

There is currently no support for history

Signed-off-by: Tim Ward <timothyjward@apache.org>