Full resource selectors (#599)
Resource Selector update

This commit enhances Resource Selectors to allow them to more easily seelct multiple providers, either of different models or with different value restrictions, and to select additional resource values that should be included with selected providers. The previous JSON syntax is still supported as a "compact resource selector" which has the same behaviour as before. The aim is to ensure full backward compatibility for existing selectors, although not binary compatibility for bundles using the API package.

The Resource Selector API now uses records throughout to ensure that the selectors are immutable, while remaining easy to serialize and deserialize.

Signed-off-by: Tim Ward <timothyjward@apache.org>