Update sensiNact GeoJSON to use Record types
This commit updates the GeoJSON types to be Java records. This significantly improves their use by making them immutable, avoiding the need to copy them repeatedly for use in notifications. The inheritance hierarchy has been replaced by interfaces, and the types are sealed to avoid unexpected extensions beyond the GeoJSON specification.

As an extra bonus, we now have proper support for *empty* GeoJSON types (which can be tested with GeoJsonObject#isEmpty()) using empty arrays in serialization, as described in the GeoJSON specification.

Signed-off-by: Tim Ward <timothyjward@apache.org>