Add core support for Linked Providers
This commit completes the support for Linked Providers in the core. This involves implementing the necessary code in the ModelNexus, adding notification support, and updating APIs/implementations to optionally include linked provider information in snapshots. As part of this change I have also properly set up the generics in the snapshots so that they don't require warning suppression.

Signed-off-by: Tim Ward <timothyjward@apache.org>