Add northbound support for Linked providers
This commit updates the northbound API, including the SensiNactSession and the Query Handler to support Linked providers. This update ensures security by applying authorization checks to each linked provider to ensure that they are visible to the caller, it also improves the efficiency of snapshotting individual providers, and ensures that all calls use the Session authorizations correctly.

Signed-off-by: Tim Ward <timothyjward@apache.org>